+++ internal/config/src/jest/get-module-map.js
@@ -1,29 +1,33 @@
 const getPackages = require('get-monorepo-packages')
-
+const path = require('path')
 // do not map modules in CI to catch any package install bugs (slower)... not in use ATM
 const doNotMapPackages = process.env.JEST_SKIP_PACKAGE_MAP === 'true'
 
 /**
- * Allows ts-jest to dynamically resolve packages so "build"
+ * Create module mapping that resolve packages for ts-jest so typescript compilation happens in-memory
  */
-const getJestModuleMap = (
-  packageRoot = '../../',
-  skipPackageMap = doNotMapPackages
-) => {
-  // get listing of packages in the mono repo
-  const createLocation = (name) => {
-    return `<rootDir>/./${name}/src/$1`
+const getJestModuleMap = ({ skipPackageMap = doNotMapPackages } = {}) => {
+  /**
+   * @param location - e.g. "packages/core"
+   */
+  const createPackageMappedPath = (location) => {
+    const base = path.basename(location) // get base folder name of a package e.g. "core" or "browser"
+    return `<rootDir>/../${base}/src/$1`
   }
-  const moduleNameMapper = getPackages(packageRoot).reduce(
+
+  // for the sake of getPackages working correctly during a project-wide test run, the working directory must be hardcoded to the root
+  const packageRoot = global.JEST_ROOT_CONFIG ? '.' : '../../'
+  const packages = getPackages(packageRoot)
+  const moduleNameMapper = packages.reduce(
     (acc, el) => ({
       ...acc,
-      [`${el.package.name}(.*)$`]: createLocation(el.location),
+      [`${el.package.name}(.*)$`]: createPackageMappedPath(el.location),
     }),
     {}
   )
 
   return {
-    '@/(.+)': '<rootdir>/../../src/$1',
+    '@/(.+)': '<rootDir>/src/$1',
     ...(skipPackageMap ? {} : moduleNameMapper),
   }
 }