+++ internal/test-helpers/.eslintrc.js
@@ -1,4 +1,7 @@
 /** @type { import('eslint').Linter.Config } */
 module.exports = {
   extends: ['../../.eslintrc'],
+  env: {
+    node: true,
+  },
 }