+++ packages/browser/src/browser/standalone-analytics.ts
@@ -2,7 +2,11 @@ import { Analytics, InitOptions } from '../core/analytics'
 import { AnalyticsBrowser } from '.'
 import { embeddedWriteKey } from '../lib/embedded-write-key'
 
-export type AnalyticsSnippet = Analytics & {
+export interface AnalyticsSnippet extends AnalyticsStandalone {
+  load: (writeKey: string, options?: InitOptions) => void
+}
+
+export interface AnalyticsStandalone extends Analytics {
   _loadOptions?: InitOptions
   _writeKey?: string
   _cdn?: string
@@ -63,5 +67,8 @@ export async function install(): Promise<void> {
     return
   }
 
-  window.analytics = await AnalyticsBrowser.standalone(writeKey, options)
+  window.analytics = (await AnalyticsBrowser.standalone(
+    writeKey,
+    options
+  )) as AnalyticsSnippet
 }