+++ packages/core/src/analytics/dispatch.ts
@@ -2,16 +2,22 @@ import { CoreContext } from '../context'
 import { CoreSegmentEvent, Callback } from '../events/interfaces'
 import { EventQueue } from '../queue/event-queue'
 import { isOffline } from '../connection'
-import { Emitter } from '../emitter'
 import { invokeCallback } from '../callback'
+import { Emitter } from '../emitter'
 
-type DispatchOptions = {
+export type DispatchOptions = {
   timeout?: number
   debug?: boolean
   callback?: Callback
   retryQueue?: boolean
 }
 
+/* The amount of time in ms to wait before invoking the callback. */
+export const getDelay = (startTimeInEpochMS: number, timeoutInMS?: number) => {
+  const elapsedTime = Date.now() - startTimeInEpochMS
+  // increasing the timeout increases the delay by almost the same amount -- this is weird legacy behavior.
+  return Math.max((timeoutInMS ?? 300) - elapsedTime, 0)
+}
 /**
  * Push an event into the dispatch queue and invoke any callbacks.
  *
@@ -27,7 +33,8 @@ export async function dispatch(
   options?: DispatchOptions
 ): Promise<CoreContext> {
   const ctx = new CoreContext(event)
-  emitter.emit('message_dispatch_pending', ctx) // TODO: inspectorHost.triggered?.(ctx as any)
+  emitter.emit('dispatch_pending', ctx) // This is just for inspector host
+  // TODO: inspectorHost.triggered?.(ctx as any)
 
   if (isOffline() && !options?.retryQueue) {
     return ctx
@@ -40,15 +47,12 @@ export async function dispatch(
   } else {
     dispatched = await queue.dispatch(ctx)
   }
-  const elapsedTime = Date.now() - startTime
-  const timeoutInMs = options?.timeout
 
   if (options?.callback) {
     dispatched = await invokeCallback(
       dispatched,
       options.callback,
-      Math.max((timeoutInMs ?? 300) - elapsedTime, 0),
-      timeoutInMs
+      getDelay(startTime, options.timeout)
     )
   }
   if (options?.debug) {